Fix Internal Server Error when creating Orders from Orders report (BO)

* Fix Internal Server Error when creating Orders from Orders report (BO)


Approved-by: Mark Jessett

+++ grails-app/services/uk/co/wonderlane/wlpos/OrderService.groovy
@@ -719,7 +719,7 @@ class OrderService extends MySqlDal  {
             productList.setEndDate(null);
         }
 
-        productList.setSupplierId(rs.getString("supplierId"));
+        productList.setSupplierId(rs.getString("supplierId") as Integer);
         if (rs.wasNull()) {
             productList.setSupplierId(null);
         }
